+++ lib/src/components/store.dart
@@ -4,12 +4,12 @@ import 'package:dash_kit_core/dash_kit_core.dart';
 class Store<S extends GlobalState> extends async_redux.Store<S> {
   Store({
     required S initialState,
-    List<ActionObserver>? actionObservers,
-    List<StateObserver>? stateObservers,
-    Persistor? persistor,
+    List<ActionObserver<S>>? actionObservers,
+    List<StateObserver<S>>? stateObservers,
+    Persistor<S>? persistor,
     ModelObserver? modelObserver,
-    ErrorObserver? errorObserver,
-    WrapError? wrapError,
+    ErrorObserver<S>? errorObserver,
+    WrapError<S>? wrapError,
   }) : super(
           initialState: initialState,
           actionObservers: actionObservers,