+++ lib/src/loadable/pagination_state.dart
@@ -1,0 +1,8 @@
+enum PaginationState {
+  idle,
+  empty,
+  loading,
+  loadingPage,
+  error,
+  errorLoadingPage,
+}