+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [3.0.0-dev.8]
+
+- Added example
+
+## [3.0.0-dev.7]
+
+- Renamed parameter in PaginatedList update function
+
+## [3.0.0-dev.6]
+
+- Updated async_redux to version 12.0.0
+- Added loadable widgets
+
 ## [3.0.0-dev.5]
 
 - Fixed an issue with types